requirements.txt
python-darkdetect python python-pyqt5
